Enum isa::Instruction

pub enum Instruction {

    InvalidInstruction,
    LOAD,
    LOADN,
    LOADI,
    STORE,
    STOREN,
    STOREI,
    MOV,
    INPUT,
    OUTPUT,
    OUTCHAR,
    INCHAR,
    SOUND,
    ADD,
    ADDC,
    SUB,
    SUBC,
    MUL,
    DIV,
    INC,
    DEC,
    MOD,
    AND,
    OR,
    XOR,
    NOT,
    SHIFTL0,
    SHIFTL1,
    SHIFTR0,
    SHIFTR1,
    ROTL,
    ROTR,
    CMP,
    JMP,
    JEQ,
    JNE,
    JZ,
    JNZ,
    JC,
    JNC,
    JGR,
    JLE,
    JEG,
    JEL,
    JOV,
    JNO,
    JDZ,
    JN,
    CALL,
    CEQ,
    CNE,
    CZ,
    CNZ,
    CC,
    CNC,
    CGR,
    CLE,
    CEG,
    CEL,
    COV,
    CNO,
    CDZ,
    CN,
    RTS,
    RTI,
    PUSH,
    POP,
    NOP,
    HALT,
    CLEARC,
    SETC,
    BREAKP,

}

Conjunto de instruções presentes na Arquitetura do Processador ICMC.

Variants

InvalidInstruction

Instrução inválida. Utilizada apenas para sinalização.

Máscara

XXXXXXXXXXXXXXXX

LOAD

Carrega o valor da memória presente no endereço END para o registrador Rx.

Operação

Rx ← MEM(END)

Uso

LOAD Rx, END

Exemplo

LOAD R3, 0xff00

Máscara

110000----------

LOADN

Carrega o valor NR no registrador Rx.

Operação

Rx ← NR

Uso

LOADN Rx, #NR

Exemplo

LOADN R3, #0xff00

Máscara

111000----------

LOADI

Carrega o valor da memória presente no endereço armazenado em Ry para o registrador Rx.

Operação

Rx ← MEM(Ry)

Uso

LOADI Rx, Ry

Exemplo

LOADI R3, R0

Máscara

111100----------

STORE

Salva no endereço END da memória o valor presente no registrador Rx.

Operação

MEM(END) ← Rx

Uso

STORE END, Rx

Exemplo

STORE 0x00ff, R3 

Máscara

110001----------

STOREN

Salva no endereço END da memória o valor NR.

Operação

MEM(END) ← NR

Uso

STOREN END, #NR

Exemplo

STOREN 0x00ff, #0b10100 

Máscara

111001----------

STOREI

Salva, na memória, no endereço armazenado em Rx, o valor presente no registrador Ry.

Operação

MEM(Rx) ← Ry

Uso

STOREI Rx, Ry

Exemplo

STOREI R3, R0 

Máscara

111101----------

MOV

Move, para um registrador Rx ou para o SP, o valor presente em outro registrador.

Operação

Rx ← Ry ou
Rx ← SP ou
SP ← Rx

Uso

MOV Rx, Ry
MOV Rx, SP
MOV SP, Rx

Exemplo

MOV R3, R0 
MOV R3, SP 
MOV SP, R0 

Máscara

110011----------

INPUT

Máscara

111110----------

OUTPUT

Máscara

111111----------

OUTCHAR

Imprime na tela do processador um char mapeado de um arquivo charmap. O código do pixelmap que representa o desenho do char está codificado no low-byte do registrador Rx, enquanto que sua cor se encontra no high-byte. A posição do char é armazenada no registrador Ry.

Cores

As cores mapeadas atualmente com seus respectivos códigos são:

⠀⠀ White — 0

⠀⠀ Brown — 256

⠀⠀ Green — 512

⠀⠀ Olive — 768

⠀⠀ Navy — 1024

⠀⠀ Purple — 1280

⠀⠀ Teal — 1536

⠀⠀ Silver — 1792

⠀⠀ Gray — 2048

⠀⠀ Red — 2304

⠀⠀ Lime — 2560

⠀⠀ Yellow — 2816

⠀⠀ Blue — 3072

⠀⠀ Fuchsia — 3328

⠀⠀ Aqua — 3584

⠀⠀ Black — 3840

Para imprimir o caracter colorido, basta somar o código do char ao código da cor.

Exemplo

A — 37 (código da letra A) + 3072 (código da cor azul).

Operação

VÍDEO(Ry) ← CHAR(Rx)

Uso

OUTCHAR Rx, Ry

Exemplo

OUTCHAR R1, R0

Máscara

110010----------

INCHAR

Máscara

110101----------

SOUND

Máscara

110100----------

ADD

Realiza a soma dos valores presentes nos registradores Ry e Rz, guardando o resultado no registrador Rx.

Operação

Rx ← Ry + Rz

Uso

ADD Rx, Ry, Rz

Exemplo

ADD R3, R0, R7 

Máscara

100000---------0

ADDC

Realiza a soma dos valores presentes nos registradores Ry e Rz mais o carry (C), guardando o resultado no registrador Rx.

Operação

Rx ← Ry + Rz + C

Uso

ADDC Rx, Ry, Rz

Exemplo

ADDC R3, R0, R7 

Máscara

100000---------1

SUB

Realiza a subtração dos valores presentes nos registradores Ry e Rz, guardando o resultado no registrador Rx.

Operação

Rx ← Ry - Rz

Uso

SUB Rx, Ry, Rz

Exemplo

SUB R3, R0, R7 

Máscara

100001---------0

SUBC

Realiza a subtração dos valores presentes nos registradores Ry e Rz, guardando no registrador Rx o resultado somado com o carry (C).

Operação

Rx ← Ry - Rz + C

Uso

SUBC Rx, Ry, Rz

Exemplo

SUBC R3, R0, R7 

Máscara

100001---------1

MUL

Realiza a multiplicação dos valores presentes nos registradores Ry e Rz, guardando o resultado no registrador Rx.

Operação

Rx ← Ry * Rz

Uso

MUL Rx, Ry, Rz

Exemplo

MUL R3, R0, R7 

Máscara

100010---------0

DIV

Realiza a divisão de Ry por Rz, guardando o resultado no registrador Rx.

Operação

Rx ← Ry / Rz

Uso

DIV Rx, Ry, Rz

Exemplo

DIV R3, R0, R7 

Máscara

100011---------0

INC

Incrementa em uma unidade o registrador Rx.

Operação

Rx ← Rx + 1

Uso

INC Rx

Exemplo

INC R3 

Máscara

100100---0------

DEC

Decrementa em uma unidade o registrador Rx.

Operação

Rx ← Rx - 1

Uso

DEC Rx

Exemplo

DEC R3 

Máscara

100100---1------

MOD

Realiza a operação de módulo entre os registradores Ry e Rz e salva o resultado no registrador Rx.

Operação

Rx ← Ry % Rz

Uso

MOD Rx, Ry, Rz

Exemplo

MOD R3, R2, R5 

Máscara

100101----------

AND

Realiza a operação AND entre os registradores Ry e Rz e salva o resultado no registrador Rx.

Operação

Rx ← Ry & Rz

Uso

AND Rx, Ry, Rz

Exemplo

AND R3, R2, R5 

Máscara

010010----------

OR

Realiza a operação OR entre os registradores Ry e Rz e salva o resultado no registrador Rx.

Operação

Rx ← Ry | Rz

Uso

OR Rx, Ry, Rz

Exemplo

OR R3, R2, R5 

Máscara

010011----------

XOR

Realiza a operação XOR entre os registradores Ry e Rz e salva o resultado no registrador Rx.

Operação

Rx ← Ry ^ Rz

Uso

XOR Rx, Ry, Rz

Exemplo

XOR R3, R2, R5 

Máscara

010100----------

NOT

Realiza a operação NOT no registrador Ry e salva o resultado no registrador Rx.

Operação

Rx ← !Ry

Uso

NOT Rx, Ry

Exemplo

NOT R3, R2 

Máscara

010101----------

SHIFTL0

Esta operação desliza os bits para a esquerda N vezes e os bits que transbordam a extremidade esquerda desaparecem. Os espaços na direita são preenchidos com 0.

Operação

Rx ← Rx << N

1 0 1 0'0 1 1 1   
 ╱ ╱ ╱ ╱ ╱ ╱ ╱
0 1 0 0'1 1 1 0  

Uso

SHIFTL0 Rx, N

Exemplo

SHIFTL0 R7, 9

Máscara

010000---000----

SHIFTL1

Esta operação desliza os bits para a esquerda N vezes e os bits que transbordam a extremidade esquerda desaparecem. Os espaços na direita são preenchidos com 1.

Operação

Rx ← !(!(Rx) << N)

0 0 1 0'0 1 1 1   
 ╱ ╱ ╱ ╱ ╱ ╱ ╱
0 1 0 0'1 1 1 1  

Uso

SHIFTL1 Rx, N

Exemplo

SHIFTL1 R7, 9

Máscara

010000---001----

SHIFTR0

Esta operação desliza os bits para a direita N vezes e os bits que transbordam a extremidade direita desaparecem. Os espaços na esquerda são preenchidos com 0.

Operação

Rx ← Rx >> N

0 0 1 0'0 1 1 1  
 ╲ ╲ ╲ ╲ ╲ ╲ ╲    
0 0 0 1'0 0 1 1  

Uso

SHIFTR0 Rx, N

Exemplo

SHIFTR0 R7, 9

Máscara

010000---010----

SHIFTR1

Esta operação desliza os bits para a direita N vezes e os bits que transbordam a extremidade direita desaparecem. Os espaços na esquerda são preenchidos com 1.

Operação

Rx ← !(!(Rx) >> N)

0 0 1 0'0 1 1 1  
 ╲ ╲ ╲ ╲ ╲ ╲ ╲    
0 0 0 1'0 0 1 1  

Uso

SHIFTR1 Rx, N

Exemplo

SHIFTR1 R7, 9

Máscara

010000---011----

ROTL

Esta operação gira os bits para a esquerda N vezes e os bits que transbordam para a extremidade esquerda são reintroduzidos no lado direito.

Operação

Rx ← (Rx << N) | (Rx >> (BITS_ADDRESS - N))

╭─────────────────╮  
│ 0 0 1 0'0 1 1 1 │  
╰─╯╱ ╱ ╱ ╱ ╱ ╱ ╱╭─╯
  0 1 0 0'1 1 1 0  

Uso

ROTL Rx, N

Exemplo

ROTL R6, 2

Máscara

010000---10-----

ROTR

Esta operação gira os bits para a direita N vezes e os bits que transbordam para a extremidade direita são reintroduzidos no lado esquerdo.

Operação

Rx ← (Rx >> N) | (Rx << (BITS_ADDRESS - N))

╭─────────────────╮  
│ 0 0 1 0'0 1 1 1 │  
╰─╮╲ ╲ ╲ ╲ ╲ ╲ ╲╰─╯    
  1 0 0 1'0 0 1 1  

Uso

ROTL Rx, N

Exemplo

ROTL R6, 2

Máscara

010000---11-----

CMP

Compara os valores dos registradores Rx e Ry e atualiza o flag register (FR) de acordo com o resultado.

Operação

FR ← COND

Uso

CMP Rx, Ry

Exemplo

CMP R3, R2 

Máscara

010110----------

JMP

Pula para o endereço END da memória.

Operação

PC ← END

Uso

JMP END 

Exemplo

JMP 0x00ff 

Máscara

0000100000------

JEQ

Pula para o endereço END da memória se o bit FlagIndex::EQUAL do flag register estiver setado.

Operação

PC ← END

Uso

JEQ END 

Exemplo

JEQ 0x00ff 

Máscara

0000100001------

JNE

Pula para o endereço END da memória se o bit FlagIndex::EQUAL do flag register não estiver setado.

Operação

PC ← END

Uso

JNE END 

Exemplo

JNE 0x00ff 

Máscara

0000100010------

JZ

Pula para o endereço END da memória se o bit FlagIndex::ZERO do flag register estiver setado.

Operação

PC ← END

Uso

JZ END 

Exemplo

JZ 0x00ff 

Máscara

0000100011------

JNZ

Pula para o endereço END da memória se o bit FlagIndex::ZERO do flag register não estiver setado.

Operação

PC ← END

Uso

JNZ END 

Exemplo

JNZ 0x00ff 

Máscara

0000100100------

JC

Pula para o endereço END da memória se o bit FlagIndex::CARRY do flag register estiver setado.

Operação

PC ← END

Uso

JC END 

Exemplo

JC 0x00ff 

Máscara

0000100101------

JNC

Pula para o endereço END da memória se o bit FlagIndex::CARRY do flag register não estiver setado.

Operação

PC ← END

Uso

JNC END 

Exemplo

JNC 0x00ff 

Máscara

0000100110------

JGR

Pula para o endereço END da memória se o bit FlagIndex::GREATER do flag register estiver setado.

Operação

PC ← END

Uso

JGR END 

Exemplo

JGR 0x00ff 

Máscara

0000100111------

JLE

Pula para o endereço END da memória se o bit FlagIndex::LESSER do flag register estiver setado.

Operação

PC ← END

Uso

JLE END 

Exemplo

JLE 0x00ff 

Máscara

0000101000------

JEG

Pula para o endereço END da memória se algum dos bits FlagIndex::GREATER ou FlagIndex::EQUAL do flag register estiver setado.

Operação

PC ← END

Uso

JEG END 

Exemplo

JEG 0x00ff 

Máscara

0000101001------

JEL

Pula para o endereço END da memória se algum dos bits FlagIndex::LESSER ou FlagIndex::EQUAL do flag register estiver setado.

Operação

PC ← END

Uso

JEL END 

Exemplo

JEL 0x00ff 

Máscara

0000101010------

JOV

Pula para o endereço END da memória se o bit FlagIndex::ARITHMETIC_OVERFLOW do flag register estiver setado.

Operação

PC ← END

Uso

JOV END 

Exemplo

JOV 0x00ff 

Máscara

0000101011------

JNO

Pula para o endereço END da memória se o bit FlagIndex::ARITHMETIC_OVERFLOW do flag register não estiver setado.

Operação

PC ← END

Uso

JNO END 

Exemplo

JNO 0x00ff 

Máscara

0000101100------

JDZ

Pula para o endereço END da memória se o bit FlagIndex::DIV_BY_ZERO do flag register estiver setado.

Operação

PC ← END

Uso

JDZ END 

Exemplo

JDZ 0x00ff 

Máscara

0000101101------

JN

Pula para o endereço END da memória se o bit FlagIndex::NEGATIVE do flag register estiver setado.

Operação

PC ← END

Uso

JN END 

Exemplo

JN 0x00ff 

Máscara

0000101110------

CALL

Salva o valor atual do PC na stack e pula para o endereço do procedimento informado.

Operação

MEM(SP) ← PC
PC ← END
SP ← SP - 1

Uso

CALL END

Exemplo

CALL 0x003C

Máscara

0000110000------

CEQ

Salva o valor atual do PC na stack e pula para o endereço do procedimento informado se o bit FlagIndex::EQUAL do flag register estiver setado.

Operação

MEM(SP) ← PC
PC ← END
SP ← SP - 1

Uso

CEQ END

Exemplo

CEQ 0x003C

Máscara

0000110001------

CNE

Salva o valor atual do PC na stack e pula para o endereço do procedimento informado se o bit FlagIndex::EQUAL do flag register não estiver setado.

Operação

MEM(SP) ← PC
PC ← END
SP ← SP - 1

Uso

CNE END

Exemplo

CNE 0x003C

Máscara

0000110010------

CZ

Salva o valor atual do PC na stack e pula para o endereço do procedimento informado se o bit FlagIndex::ZERO do flag register estiver setado.

Operação

MEM(SP) ← PC
PC ← END
SP ← SP - 1

Uso

CZ END

Exemplo

CZ 0x003C

Máscara

0000110011------

CNZ

Salva o valor atual do PC na stack e pula para o endereço do procedimento informado se o bit FlagIndex::ZERO do flag register não estiver setado.

Operação

MEM(SP) ← PC
PC ← END
SP ← SP - 1

Uso

CNZ END

Exemplo

CNZ 0x003C

Máscara

0000110100------

CC

Salva o valor atual do PC na stack e pula para o endereço do procedimento informado se o bit FlagIndex::CARRY do flag register estiver setado.

Operação

MEM(SP) ← PC
PC ← END
SP ← SP - 1

Uso

CC END

Exemplo

CC 0x003C

Máscara

0000110101------

CNC

Salva o valor atual do PC na stack e pula para o endereço do procedimento informado se o bit FlagIndex::CARRY do flag register não estiver setado.

Operação

MEM(SP) ← PC
PC ← END
SP ← SP - 1

Uso

CNC END

Exemplo

CNC 0x003C

Máscara

0000110110------

CGR

Salva o valor atual do PC na stack e pula para o endereço do procedimento informado se o bit FlagIndex::GREATER do flag register estiver setado.

Operação

MEM(SP) ← PC
PC ← END
SP ← SP - 1

Uso

CGR END

Exemplo

CGR 0x003C

Máscara

0000110111------

CLE

Salva o valor atual do PC na stack e pula para o endereço do procedimento informado se o bit FlagIndex::LESSER do flag register estiver setado.

Operação

MEM(SP) ← PC
PC ← END
SP ← SP - 1

Uso

CLE END

Exemplo

CLE 0x003C

Máscara

0000111000------

CEG

Salva o valor atual do PC na stack e pula para o endereço do procedimento informado se algum dos bits FlagIndex::EQUAL ou FlagIndex::GREATER do flag register estiver setado

Operação

MEM(SP) ← PC
PC ← END
SP ← SP - 1

Uso

CEG END

Exemplo

CEG 0x003C

Máscara

0000111001------

CEL

Salva o valor atual do PC na stack e pula para o endereço do procedimento informado se algum dos bits FlagIndex::EQUAL ou FlagIndex::LESSER do flag register estiver setado

Operação

MEM(SP) ← PC
PC ← END
SP ← SP - 1

Uso

CEL END

Exemplo

CEL 0x003C

Máscara

0000111010------

COV

Salva o valor atual do PC na stack e pula para o endereço do procedimento informado se o bit FlagIndex::ARITHMETIC_OVERFLOW do flag register estiver setado.

Operação

MEM(SP) ← PC
PC ← END
SP ← SP - 1

Uso

COV END

Exemplo

COV 0x003C

Máscara

0000111011------

CNO

Salva o valor atual do PC na stack e pula para o endereço do procedimento informado se o bit FlagIndex::ARITHMETIC_OVERFLOW do flag register não estiver setado.

Operação

MEM(SP) ← PC
PC ← END
SP ← SP - 1

Uso

CNO END

Exemplo

CNO 0x003C

Máscara

0000111100------

CDZ

Salva o valor atual do PC na stack e pula para o endereço do procedimento informado se o bit FlagIndex::DIV_BY_ZERO do flag register estiver setado.

Operação

MEM(SP) ← PC
PC ← END
SP ← SP - 1

Uso

CDZ END

Exemplo

CDZ 0x003C

Máscara

0000111101------

CN

Salva o valor atual do PC na stack e pula para o endereço do procedimento informado se o bit FlagIndex::NEGATIVE do flag register estiver setado.

Operação

MEM(SP) ← PC
PC ← END
SP ← SP - 1

Uso

CN END

Exemplo

CN 0x003C

Máscara

0000111110------

RTS

Altera o valor do PC para o último valor salvo na stack somado de 1.

Operação

SP ← SP + 1
PC ← MEM(SP)
PC ← PC + 1

Uso

RTS

Exemplo

RTS

Máscara

000100---------0

RTI

Altera o valor do PC para o último valor salvo na stack.

Operação

SP ← SP + 1
PC ← MEM(SP)

Uso

RTI

Exemplo

RTI

Máscara

000100---------1

PUSH

Salva na stack o conteúdo de um registrador ou do flag register.

Operação

MEM(SP) ← Rx
SP ← SP - 1 ou
MEM(SP) ← FR
SP ← SP - 1

Uso

PUSH Rx  
PUSH FR

Exemplo

PUSH R5
PUSH FR

Máscara

000101----------

POP

Recupera da stack o conteúdo de um registrador ou do flag register.

Operação

SP ← SP + 1
Rx ← MEM(SP) ou
SP ← SP + 1
FR ← MEM(SP)

Uso

POP Rx  
POP FR

Exemplo

POP R5
POP FR

Máscara

000110----------

NOP

Sem operação. Serve apenas para consumir tempo.

Operação

Nenhuma

Uso

NOP

Exemplo

NOP

Máscara

000000----------

HALT

Para a execução do programa.

Operação

Para o processador

Uso

HALT

Exemplo

HALT

Máscara

001111----------

CLEARC

Limpa o bit FlagIndex::CARRY do flag register.

Operação

C ← 0

Uso

CLEARC

Exemplo

CLEARC

Máscara

0010000---------

SETC

Seta o bit FlagIndex::CARRY do flag register.

Operação

C ← 1

Uso

SETC

Exemplo

SETC

Máscara

0010001---------

BREAKP

Gera um breakpoint no código, forçando o simulador a entrar no modo debug.

Operação

Nenhuma operação lógica no processador

Uso

BREAKP

Exemplo

BREAKP

Máscara

001110----------

Implementations

impl Instruction

pub fn vec() -> Vec<Instruction>

pub fn opcode(&self) -> MemoryCell

Retorna o OPCODE da instrução.

pub fn mask(&self) -> &str

Retorna a máscara da instrução.

pub fn from_str(s: &str) -> Option<Self>

pub fn bits(&self, r: RangeInclusive<usize>) -> MemoryCell

pub fn get_instruction(value: MemoryCell) -> Instruction

Retorna qual Instruction está presente no argumento value. Se a instrução for inválida, irá retornar Instruction::InvalidInstruction.

Exemplo

use isa::*;

let mem = 0b1100001000100011; // LOAD
assert_eq!(Instruction::LOAD, Instruction::get_instruction(mem));